Validate that the API key is present, as this is the standard for those types of packages.
For example, if you import mongo/postgress and you don't have the right credential in place the spring boto server will throw an error during start time. This is somewhat the DNA in spring boot that if you chose to bring a dependency you have to play by its rules.